backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
sqlmodel
google-genai
python-dotenv
aiofiles
